refactor: Split apps/api-runtime into - apps/api-runtime (Mongo dependent) and libs/dependency-analyzer (mongo independent)
