+++ apps/api-runtime/src/test/java/com/akto/dependency/TestDependencyAnalyser.java
@@ -1,5 +1,6 @@
 package com.akto.dependency;
 
+import com.akto.DependencyAnalyserHelper;
 import com.akto.MongoBasedTest;
 import com.akto.dao.*;
 import com.akto.dao.context.Context;
@@ -248,23 +249,23 @@ public class TestDependencyAnalyser extends MongoBasedTest {
         templateURLToMethods.put(RuntimeUtil.createUrlTemplate("api/hotel/INTEGER", URLMethods.Method.POST), null);
         apiCatalog.setTemplateURLToMethods(templateURLToMethods);
         dbState.put(1000, apiCatalog);
-        DependencyAnalyser dependencyAnalyser = new DependencyAnalyser(dbState);
-        dependencyAnalyser.nodes = nodes;
+        DependencyAnalyserHelper dependencyAnalyserHelper = new DependencyAnalyserHelper(dbState);
+        dependencyAnalyserHelper.setNode(nodes);
 
-        assertEquals(6, dependencyAnalyser.nodes.size());
-        dependencyAnalyser.mergeNodes();
-        assertEquals(3, dependencyAnalyser.nodes.size());
+        assertEquals(6, dependencyAnalyserHelper.getNodes().size());
+        dependencyAnalyserHelper.mergeNodes();
+        assertEquals(3, dependencyAnalyserHelper.getNodes().size());
 
 
-        DependencyNode dependencyNode1New = dependencyAnalyser.nodes.get(Objects.hash("1000", "api/books/INTEGER", "GET","1000", "/api/cars", "POST"));
+        DependencyNode dependencyNode1New = dependencyAnalyserHelper.getNodes().get(Objects.hash("1000", "api/books/INTEGER", "GET","1000", "/api/cars", "POST"));
         assertEquals(1, dependencyNode1New.getParamInfos().size());
         assertEquals(3, dependencyNode1New.getParamInfos().get(0).getCount());
 
-        DependencyNode dependencyNode2New = dependencyAnalyser.nodes.get(Objects.hash("1000", "api/toys/INTEGER", "GET","1000", "/api/bus", "POST"));
+        DependencyNode dependencyNode2New = dependencyAnalyserHelper.getNodes().get(Objects.hash("1000", "api/toys/INTEGER", "GET","1000", "/api/bus", "POST"));
         assertEquals(1, dependencyNode2New.getParamInfos().size());
         assertEquals(2, dependencyNode2New.getParamInfos().get(0).getCount());
 
-        DependencyNode dependencyNode3New = dependencyAnalyser.nodes.get(Objects.hash("1000", "api/food/INTEGER", "GET","1000", "api/hotel/INTEGER", "POST"));
+        DependencyNode dependencyNode3New = dependencyAnalyserHelper.getNodes().get(Objects.hash("1000", "api/food/INTEGER", "GET","1000", "api/hotel/INTEGER", "POST"));
         assertEquals(1, dependencyNode3New.getParamInfos().size());
         assertEquals(4, dependencyNode3New.getParamInfos().get(0).getCount());
     }
