+++ libs/dto/src/main/java/com/akto/util/store/HashSetStore.java
@@ -1,4 +1,4 @@
-package com.akto.dependency.store;
+package com.akto.util.store;
 
 import java.util.HashSet;
 import java.util.Set;
