+++ libs/dto/src/main/java/com/akto/util/store/Store.java
@@ -1,4 +1,4 @@
-package com.akto.dependency.store;
+package com.akto.util.store;
 
 public abstract class Store {
